Converted package bdede version BLP_LIB_BDE_BDE_2.5.0 from clearcase

+++ groups/bde/bdede/bdede_quotedprintableencoder.h
@@ -721,7 +721,7 @@ int bdede_QuotedPrintableEncoder::maxLineLength() const
 inline
 int bdede_QuotedPrintableEncoder::numOutputPending() const
 {
-    return d_outBuf.size();
+    return static_cast<int>(d_outBuf.size());
 }
 
 inline
